Adverb for new column-pair mapping functions

as_colpair_mapper() takes a function f and returns a new function that applies f to each pair of columns in a data frame.

Usage

as_colpair_mapper(
  f,
  eval_f = TRUE,
  class = TRUE,
  default_diagonal = NA,
  default_quiet = FALSE,
  ...
)

Arguments

f

A function that takes two or more arguments.

eval_f

Boolean. If set to FALSE, f will appear in the output function by name rather than as its body and arguments. Default is TRUE. See details.

class

Boolean. If TRUE (the default), the data frames returned by the output function will inherit a class that starts on "pairmaps_df_", followed by the name of f.

default_diagonal, default_quiet

Defaults for the .diagonal and .quiet arguments of the output function. By default of the present function, these will be NA and FALSE, respectively.

...

These dots must be empty.

Value

A function. See vary() and covary() for examples of functions made by as_colpair_mapper().

Details

Setting eval_f to FALSE can be helpful if you copy and paste the resulting function into a script. In this case, it makes sense to refer to one particular function by name, rather than copying and pasting a lengthy function definition within your new function definition, and manually updating it when the mapped function changes.

However, this should only be considered if the value of f is completely unambiguous, because such ambiguity might lead to a very subtle bug. See vignette("using-pairmaps"), section Assignment and copying.

References

Wickham, H. (2019). Advanced R (Second Edition), CRC Press/Taylor and Francis Group. https://adv-r.hadley.nz/index.html.

See also

Examples

# From `?corrr::colpair_map()` -- a function that
# extracts the p-value from a t-test:
calc_p_value <- function(vec_a, vec_b) {
  stats::t.test(vec_a, vec_b)$p.value
}

# Derive a new mapping function:
p_value_map <- as_colpair_mapper(f = calc_p_value)

# Below are three near-equivalent calls.
# The only difference: In the first two cases,
# the output tibble inherits an extra class
# that includes the name of `f`. Here, it is
# `"pairmaps_df_calc_p_value"`.

# 1. Using a ready-made mapping function:
p_value_map(.data = mtcars)
#> ℹ Applying `calc_p_value` to each column pair
#> # A tibble: 11 × 12
#>    term        mpg       cyl      disp        hp      drat        wt      qsec
#>    <chr>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 mpg   NA         9.51e-15  7.98e-11  1.03e-11  3.16e-16  1.03e-16  5.11e- 2
#>  2 cyl    9.51e-15 NA         1.77e-11  8.32e-13  2.28e- 9  9.12e-11  3.73e-35
#>  3 disp   7.98e-11  1.77e-11 NA         1.55e- 3  1.35e-11  1.29e-11  6.34e-11
#>  4 hp     1.03e-11  8.32e-13  1.55e- 3 NA         5.28e-13  4.92e-13  7.24e-12
#>  5 drat   3.16e-16  2.28e- 9  1.35e-11  5.28e-13 NA         6.02e- 2  5.91e-33
#>  6 wt     1.03e-16  9.12e-11  1.29e-11  4.92e-13  6.02e- 2 NA         7.27e-39
#>  7 qsec   5.11e- 2  3.73e-35  6.34e-11  7.24e-12  5.91e-33  7.27e-39 NA       
#>  8 vs     2.24e-18  3.50e-19  9.62e-12  3.01e-13  2.43e-33  1.33e-18  1.05e-35
#>  9 am     2.15e-18  3.07e-19  9.59e-12  3.00e-13  1.14e-33  9.09e-19  1.13e-35
#> 10 gear   3.08e-16  5.64e- 9  1.36e-11  5.36e-13  5.75e- 1  3.41e- 2  2.97e-35
#> 11 carb   1.68e-17  5.61e-11  1.24e-11  4.55e-13  1.30e- 2  2.31e- 1  1.74e-42
#> # … with 4 more variables: vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>

# 2. Directly working with the adverb:
as_colpair_mapper(f = calc_p_value)(.data = mtcars)
#> ℹ Applying `calc_p_value` to each column pair
#> # A tibble: 11 × 12
#>    term        mpg       cyl      disp        hp      drat        wt      qsec
#>    <chr>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 mpg   NA         9.51e-15  7.98e-11  1.03e-11  3.16e-16  1.03e-16  5.11e- 2
#>  2 cyl    9.51e-15 NA         1.77e-11  8.32e-13  2.28e- 9  9.12e-11  3.73e-35
#>  3 disp   7.98e-11  1.77e-11 NA         1.55e- 3  1.35e-11  1.29e-11  6.34e-11
#>  4 hp     1.03e-11  8.32e-13  1.55e- 3 NA         5.28e-13  4.92e-13  7.24e-12
#>  5 drat   3.16e-16  2.28e- 9  1.35e-11  5.28e-13 NA         6.02e- 2  5.91e-33
#>  6 wt     1.03e-16  9.12e-11  1.29e-11  4.92e-13  6.02e- 2 NA         7.27e-39
#>  7 qsec   5.11e- 2  3.73e-35  6.34e-11  7.24e-12  5.91e-33  7.27e-39 NA       
#>  8 vs     2.24e-18  3.50e-19  9.62e-12  3.01e-13  2.43e-33  1.33e-18  1.05e-35
#>  9 am     2.15e-18  3.07e-19  9.59e-12  3.00e-13  1.14e-33  9.09e-19  1.13e-35
#> 10 gear   3.08e-16  5.64e- 9  1.36e-11  5.36e-13  5.75e- 1  3.41e- 2  2.97e-35
#> 11 carb   1.68e-17  5.61e-11  1.24e-11  4.55e-13  1.30e- 2  2.31e- 1  1.74e-42
#> # … with 4 more variables: vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>

# 3. Calling the underlying function:
corrr::colpair_map(.data = mtcars, .f = calc_p_value)
#> # A tibble: 11 × 12
#>    term        mpg       cyl      disp        hp      drat        wt      qsec
#>    <chr>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 mpg   NA         9.51e-15  7.98e-11  1.03e-11  3.16e-16  1.03e-16  5.11e- 2
#>  2 cyl    9.51e-15 NA         1.77e-11  8.32e-13  2.28e- 9  9.12e-11  3.73e-35
#>  3 disp   7.98e-11  1.77e-11 NA         1.55e- 3  1.35e-11  1.29e-11  6.34e-11
#>  4 hp     1.03e-11  8.32e-13  1.55e- 3 NA         5.28e-13  4.92e-13  7.24e-12
#>  5 drat   3.16e-16  2.28e- 9  1.35e-11  5.28e-13 NA         6.02e- 2  5.91e-33
#>  6 wt     1.03e-16  9.12e-11  1.29e-11  4.92e-13  6.02e- 2 NA         7.27e-39
#>  7 qsec   5.11e- 2  3.73e-35  6.34e-11  7.24e-12  5.91e-33  7.27e-39 NA       
#>  8 vs     2.24e-18  3.50e-19  9.62e-12  3.01e-13  2.43e-33  1.33e-18  1.05e-35
#>  9 am     2.15e-18  3.07e-19  9.59e-12  3.00e-13  1.14e-33  9.09e-19  1.13e-35
#> 10 gear   3.08e-16  5.64e- 9  1.36e-11  5.36e-13  5.75e- 1  3.41e- 2  2.97e-35
#> 11 carb   1.68e-17  5.61e-11  1.24e-11  4.55e-13  1.30e- 2  2.31e- 1  1.74e-42
#> # … with 4 more variables: vs <dbl>, am <dbl>, gear <dbl>, carb <dbl>

# (In the second call, R first evaluates
# `as_colpair_mapper(f = calc_p_value)` to a
# mapping function just like `p_value_map()`,
# and then calls that new function on `mtcars`.)