Variance and covariance data frames

vary() and covary() return variance and covariance data frames instead of matrices, making them easier to explore with the corrr package.

Usage

vary(.data, ..., .diagonal = NA, .quiet = FALSE)

covary(.data, ..., .diagonal = NA, .quiet = FALSE)

Arguments

.data

Numeric vector, matrix or data frame.

...

Further arguments passed on to var() or cov(), respectively.

.diagonal

Value to which the diagonal will be set. Default is NA.

.quiet

Boolean. Set it to TRUE to suppress the message about function mapping.

Value

Both functions return cor_df tibbles, i.e., "correlation data frames". However, they don't compute correlations: This name simply reflects the origins of the cor_df class in the corrr package.

See also

Examples

vary(mtcars)
#> ℹ Applying `stats::var` to each column pair
#> # A tibble: 11 × 12
#>    term      mpg     cyl   disp      hp     drat      wt     qsec       vs
#>    <chr>   <dbl>   <dbl>  <dbl>   <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 mpg     NA     -9.17  -633.  -321.     2.20    -5.12    4.51     2.02  
#>  2 cyl     -9.17  NA      200.   102.    -0.668    1.37   -1.89    -0.730 
#>  3 disp  -633.   200.      NA   6721.   -47.1    108.    -96.1    -44.4   
#>  4 hp    -321.   102.    6721.    NA    -16.5     44.2   -86.8    -25.0   
#>  5 drat     2.20  -0.668  -47.1  -16.5   NA       -0.373   0.0871   0.119 
#>  6 wt      -5.12   1.37   108.    44.2   -0.373   NA      -0.305   -0.274 
#>  7 qsec     4.51  -1.89   -96.1  -86.8    0.0871  -0.305  NA        0.671 
#>  8 vs       2.02  -0.730  -44.4  -25.0    0.119   -0.274   0.671   NA     
#>  9 am       1.80  -0.466  -36.6   -8.32   0.190   -0.338  -0.205    0.0423
#> 10 gear     2.14  -0.649  -50.8   -6.36   0.276   -0.421  -0.280    0.0766
#> 11 carb    -5.36   1.52    79.1   83.0   -0.0784   0.676  -1.89    -0.464 
#> # … with 3 more variables: am <dbl>, gear <dbl>, carb <dbl>

covary(mtcars)
#> ℹ Applying `stats::cov` to each column pair
#> # A tibble: 11 × 12
#>    term      mpg     cyl   disp      hp     drat      wt     qsec       vs
#>    <chr>   <dbl>   <dbl>  <dbl>   <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 mpg     NA     -9.17  -633.  -321.     2.20    -5.12    4.51     2.02  
#>  2 cyl     -9.17  NA      200.   102.    -0.668    1.37   -1.89    -0.730 
#>  3 disp  -633.   200.      NA   6721.   -47.1    108.    -96.1    -44.4   
#>  4 hp    -321.   102.    6721.    NA    -16.5     44.2   -86.8    -25.0   
#>  5 drat     2.20  -0.668  -47.1  -16.5   NA       -0.373   0.0871   0.119 
#>  6 wt      -5.12   1.37   108.    44.2   -0.373   NA      -0.305   -0.274 
#>  7 qsec     4.51  -1.89   -96.1  -86.8    0.0871  -0.305  NA        0.671 
#>  8 vs       2.02  -0.730  -44.4  -25.0    0.119   -0.274   0.671   NA     
#>  9 am       1.80  -0.466  -36.6   -8.32   0.190   -0.338  -0.205    0.0423
#> 10 gear     2.14  -0.649  -50.8   -6.36   0.276   -0.421  -0.280    0.0766
#> 11 carb    -5.36   1.52    79.1   83.0   -0.0784   0.676  -1.89    -0.464 
#> # … with 3 more variables: am <dbl>, gear <dbl>, carb <dbl>